split route and controller parts from smart build
